+++ spec/support/create_test_database.rb
@@ -6,6 +6,8 @@ ActiveRecord::Schema.define version: 0 do
   create_table :campaigns, force: true do |t|
     t.datetime :start_at
     t.datetime :end_at
+    t.datetime :test_start_at
+    t.datetime :test_end_at
   end
 
   create_table :events, force: true do |t|